// src/modules/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  NEW_INFLUENCER_REVIEW = 'NEW_INFLUENCER_REVIEW',
  ACCOUNT_APPROVED = 'ACCOUNT_APPROVED',
  ACCOUNT_REJECTED = 'ACCOUNT_REJECTED',
  CAMPAIGN_SUBMITTED_FOR_REVIEW = 'CAMPAIGN_SUBMITTED_FOR_REVIEW',
  CAMPAIGN_APPROVED = 'CAMPAIGN_APPROVED',
  CAMPAIGN_REJECTED = 'CAMPAIGN_REJECTED',
  CAMPAIGN_INVITATION = 'CAMPAIGN_INVITATION',
  CAMPAIGN_PENDING_MINIMUM = 'CAMPAIGN_PENDING_MINIMUM',
  CAMPAIGN_AUTO_DISCARDED = 'CAMPAIGN_AUTO_DISCARDED',
  NEW_CAMPAIGN_APPLICATION = 'NEW_CAMPAIGN_APPLICATION',
  NEW_APPLICATION_OFFER = 'NEW_APPLICATION_OFFER',
  APPLICATION_OFFER_REJECTED = 'APPLICATION_OFFER_REJECTED',
  APPLICATION_ACCEPTED = 'APPLICATION_ACCEPTED',
  APPLICATION_REJECTED = 'APPLICATION_REJECTED',
  CAMPAIGN_STARTED = 'CAMPAIGN_STARTED',
  NEW_CONTENT_SUBMISSION = 'NEW_CONTENT_SUBMISSION',
  SUBMISSION_ACCEPTED = 'SUBMISSION_ACCEPTED',
  SUBMISSION_MODIFICATION_REQUESTED = 'SUBMISSION_MODIFICATION_REQUESTED',
}